Fixed a few issues with my applicative code, implemented python apply for the most part
I still need to fix some path issues with apply, see todo in the actual code. I learned the valuable lesson, that you can't
at all deal with types f(a->b), at least it is impossible to generate them and I see no tangible benefit in doing so. Everything
is more nice and tidy if we use f a -> f b, since we can now just take an element out of the first type. Note that the latter form
can always be gained from the other one with the (<$>) operator. I now want to let the user stay inside the f a world as long as
possible by using a typeclass-based prelude and then implementing the corresponding classes for f Int, f Bool, f String, etc.