Wrote python import code for arbitrary paths
I think I am going to push the 'capability based' layer more to the front. It will always be impossible
to write a completely segment the IO monad in a reasonable way for all intends and purposes. So I need
to write domain-based IO segmentation, so my python capabilties can do arbitrary IO. This is motivated
by the fact that I need to check whether a file exists and I don't know how and also don't want to write
insane amounts of boilerplate.