Replaced String with LText, moved to protolude, discoverd path library, moved capabilties to front
This venture has been a big fight againts compiler errors. Changing so many things at once has caused
me to rewrite most of the Python and Capabilities code. It took somewhere around 10 hours, it is not clear
whether it will pay off, but it is certain that the cost of changing it later would have been significantly
bigger. I can now really start porting more of the code.

Having only a single addContent function for the target path is I think completely sufficient for the time
being.