Wrap in a nix package(after a week of pain and suffering)
I have used the poorly documented haskell.nix nix library to make my stack project into a nix project.
The code I have produced makes it seem like this was a very easy feat, but I have a lot of time reading
the source code of that library and also gaining a better understanding of nix in order to produce that
code. I don't exactly know what I'm doing, and I have said at many points "who cares, it works" and I think
that is fine because I have better things to do. This was not fun at all, but it is necessary if I want to
adhere to my own standards of making everything perfectly reproducible. Now the kaggle library is automatically
included when I install iterbuild! At least adding new dependencies in the future will be really easy, because
I have figured out how to do that.